refactor: move the toolkit-free code out of the UI package
Importing flimkit.interactive pulled in all sixteen UI modules and tkinter,
because stitch.py imported GUI_MODE from UI.gui and then never used it. That
import is gone, so the analysis core now imports on a machine with no tk.

RoiManager holds region geometry, masks and GeoJSON and has no toolkit code
in it, so it moves to flimkit/utils/roi.py with the three patch helpers.
flim_display.py moves whole to flimkit/utils/display.py. _C and the session
NPZ helpers come out of UI/utils.py into utils/config_snapshot.py and
utils/session.py. roi_tools re-exports RoiManager and the patch helpers, so
plugin and bridge imports keep working.

The moved code is otherwise unchanged. Two function-local imports in the
session helpers are now module-level, and that is the only edit. Checked by
comparing the ASTs before and after.

test_headless_imports.py imports the core modules in a subprocess and fails
if anything under tkinter or flimkit.UI turns up in sys.modules, then imports
them again with tkinter unimportable. 73 ROI and display tests now run with
the whole UI package blocked. The first version of that test was useless: a
try/except around a blocked import swallowed it and the test passed anyway.

roi_tools.py and project_panel.py lost their comments and docstrings while
they were open. The module reference and project tree in the docs follow the
move.

780 passing.
