+++ flimkit/UI/gui.py
@@ -27,8 +27,9 @@ from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
 from matplotlib.figure import Figure
 from flimkit.UI.progress_window import ProgressWindow
 from flimkit.UI.phasor_panel import PhasorViewPanel
-from flimkit.UI import flim_display
-from flimkit.UI.roi_tools import RoiManager, RoiAnalysisPanel
+from flimkit.utils import display
+from flimkit.utils.roi import RoiManager
+from flimkit.UI.roi_tools import RoiAnalysisPanel
 from flimkit.UI.project_panel import ProjectBrowserPanel
 
 try:
@@ -1808,7 +1809,7 @@ Anthropic's Claude AI assisted with parts of the GUI implementation.
                     self._fov_preview._n_exp = int(n_exp)
             try:
                 if self._fov_preview._lifetime_map is not None and self._fov_preview._intensity_map is not None:
-                    from flimkit.UI import flim_display
+                    from flimkit.utils import display
                     intensity = self._fov_preview._intensity_map
                     lifetime = self._fov_preview._lifetime_map
                     ax_img = self._fov_preview._ax_img
@@ -1822,13 +1823,13 @@ Anthropic's Claude AI assisted with parts of the GUI implementation.
                     ax_img.set_xlabel('X (pixels)')
                     ax_img.set_ylabel('Y (pixels)')
                     cs = self._fov_preview._flim_color_scale
-                    scaled = flim_display.apply_color_scale(
+                    scaled = display.apply_color_scale(
                         lifetime,
                         vmin=cs.get('vmin'),
                         vmax=cs.get('vmax'),
                         gamma=cs.get('gamma', 1.0),
                     )
-                    cmap_obj = flim_display.get_colormap(cs.get('cmap', 'viridis'))
+                    cmap_obj = display.get_colormap(cs.get('cmap', 'viridis'))
                     cmap_obj.set_bad(color='black')
                     ax_flim.clear()
                     ax_cbar.clear()
@@ -2096,7 +2097,7 @@ Anthropic's Claude AI assisted with parts of the GUI implementation.
                         if with_scalebar:
                             self._draw_scale_bar(ax, w, h, pixel_size_um)
                         if with_annotations and self._fov_preview._roi_manager.get_all_regions():
-                            from flimkit.UI.roi_tools import get_rectangle_patch, get_ellipse_patch, get_polygon_patch
+                            from flimkit.utils.roi import get_rectangle_patch, get_ellipse_patch, get_polygon_patch
                             for region in self._fov_preview._roi_manager.get_all_regions():
                                 region_id = region['id']
                                 tool_type = region['tool']
@@ -3219,7 +3220,7 @@ Anthropic's Claude AI assisted with parts of the GUI implementation.
                 f'✓  Z-stack complete - {n_stacks} stack(s), {n_slices} slices.')
             first = next(iter(result.values())) if isinstance(result, dict) and result else None
             if first is not None:
-                from flimkit.UI.flim_display import load_zstack_display_slices
+                from flimkit.utils.display import load_zstack_display_slices
                 try:
                     slices = load_zstack_display_slices(first['group_dir'], ptu_dir=ptu_dir)
                     if slices:
