+++ flimkit_tests/tests/test_headless_imports.py
@@ -1,0 +1,64 @@
+import os
+import subprocess
+import sys
+
+CORE_MODULES = [
+    'flimkit',
+    'flimkit.configs',
+    'flimkit.dialogs',
+    'flimkit.interactive',
+    'flimkit.project',
+    'flimkit.synth',
+    'flimkit.formats',
+    'flimkit.formats.PTU.stitch',
+    'flimkit.FLIM.fitters',
+    'flimkit.FLIM.irf_tools',
+    'flimkit.phasor',
+    'flimkit.plugins',
+    'flimkit.utils.config_snapshot',
+    'flimkit.utils.display',
+    'flimkit.utils.roi',
+    'flimkit.utils.session',
+]
+
+PURITY = '''
+import sys
+import importlib
+
+for name in {modules!r}:
+    importlib.import_module(name)
+
+leaked = sorted(m for m in sys.modules if m == 'tkinter' or m.startswith('tkinter.') or m.startswith('flimkit.UI'))
+if leaked:
+    raise SystemExit('headless core pulled in the desktop frontend: ' + ', '.join(leaked))
+'''
+
+MISSING_TK = '''
+import sys
+import importlib
+import importlib.abc
+
+class Block(importlib.abc.MetaPathFinder):
+    def find_spec(self, name, path=None, target=None):
+        if name == 'tkinter' or name.startswith('tkinter.'):
+            raise ImportError('no tkinter on this machine')
+        return None
+
+sys.meta_path.insert(0, Block())
+for name in {modules!r}:
+    importlib.import_module(name)
+'''
+
+def run_probe(source):
+    env = dict(os.environ, MPLBACKEND='Agg')
+    return subprocess.run(
+        [sys.executable, '-c', source.format(modules=CORE_MODULES)],
+        capture_output=True, text=True, env=env)
+
+def test_core_does_not_import_the_desktop_frontend():
+    proc = run_probe(PURITY)
+    assert proc.returncode == 0, proc.stdout + proc.stderr
+
+def test_core_imports_on_a_machine_without_tkinter():
+    proc = run_probe(MISSING_TK)
+    assert proc.returncode == 0, proc.stderr
